app/requirements.txt
fastapi
sqlalchemy
aiokafka
python-dotenv
uvicorn
